add servicesPage and postsPage
