+++ src/routes/postsPage/postsPage.component.jsx
@@ -1,0 +1,14 @@
+import Footer from "../../components/footer/footer.component";
+import Posts from "../posts/posts.component";
+
+
+const PostsPage = () => {
+    return(
+        <>
+            <Posts />
+            <Footer />
+        </>
+    );
+}
+
+export default PostsPage;
