+++ src/routes/servicesPage/servicesPage.component.jsx
@@ -1,0 +1,13 @@
+import Footer from "../../components/footer/footer.component";
+import Services from "../services/services.component";
+
+const ServicesPage = () => {
+    return(
+        <>
+            <Services />
+            <Footer />
+        </>
+    );
+}
+
+export default ServicesPage;
